Update editor_gui.py
Fix the issue where, when line numbers are toggled off and then on again, the newly spawned text widget is not packed correctly and appears at the bottom of the frame. Forcing a redraw of the text area fixes this though I'm not sure this is ideal? Perhaps.

Formatting fixes/line number color.