Update editor_gui.py
Wow that took a while. Now line numbers scroll with the text area. I disabled the manual scrolling of the line numbers widget, so now you can't scroll the text area if your mouse is over the line numbers, which is not ideal. Maybe reenable scroll and set the yscrollcommand of the text area to the yview of the line number pane?